requirements.txt
flask==2.3.3
playwright==1.40.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
selenium==4.15.2
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
